refactor: don't introduce MessageFlagsEnum, use MessageFlags.flag
